Enhance search
Treat each word in the search string separately, and use each word
separately to refine the results.

So, artists are returned if all search words appear in their name.
(This allows first name and second name to be searched in reverse order)

Albums are returned if each word appears in the album title or in
the artist name.

Tracks are returned if each word appears in the track title, the
album title, or in the artist name.